+++ src/main/java/com/github/cpburnz/minecraft_prometheus_exporter/MinecraftCollector.java
@@ -3,17 +3,19 @@ package com.github.cpburnz.minecraft_prometheus_exporter;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import javax.annotation.Nullable;
 
 import com.mojang.authlib.GameProfile;
-
-import io.prometheus.client.Collector;
-import io.prometheus.client.GaugeMetricFamily;
-import io.prometheus.client.Histogram;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.level.Level;
+
+import io.prometheus.client.Collector;
+import io.prometheus.client.GaugeMetricFamily;
+import io.prometheus.client.Histogram;
+
 
 /**
  * This class collects stats from the Minecraft server for export.
@@ -34,32 +36,37 @@ public class MinecraftCollector extends Collector implements Collector.Describab
 	};
 
 	/**
+	 * The active dimension being timed.
+	 */
+	@Nullable
+	private String dim_tick_name;
+
+	/**
+	 * Histogram metrics for dimension tick timing.
+	 */
+	private final Histogram dim_tick_seconds;
+
+	/**
+	 * The active timer when timing a dimension tick.
+	 */
+	@Nullable
+	private Histogram.Timer dim_tick_timer;
+
+	/**
 	 * The Minecraft server.
 	 */
-	@SuppressWarnings("FieldMayBeFinal")
-	private MinecraftServer mc_server;
+	private final MinecraftServer mc_server;
 
 	/**
 	 * Histogram metrics for server tick timing.
 	 */
-	@SuppressWarnings("FieldMayBeFinal")
-	private Histogram server_tick_seconds;
+	private final Histogram server_tick_seconds;
 
 	/**
-	 * Timer used for server ticks.
+	 * The active timer when timing a server tick.
 	 */
+	@Nullable
 	private Histogram.Timer server_tick_timer;
-
-	/**
-	 * Histogram metrics for world tick timing.
-	 */
-	@SuppressWarnings("FieldMayBeFinal")
-	private Histogram world_tick_seconds;
-
-	/**
-	 * Timer used for world ticks.
-	 */
-	private Histogram.Timer world_tick_timer;
 
 	/**
 	 * Constructs the instance.
@@ -76,11 +83,11 @@ public class MinecraftCollector extends Collector implements Collector.Describab
 			.help("Stats on server tick times.")
 			.create();
 
-		this.world_tick_seconds = Histogram.build()
+		this.dim_tick_seconds = Histogram.build()
 			.buckets(TICK_BUCKETS)
-			.name("mc_world_tick_seconds")
+			.name("mc_dimension_tick_seconds")
 			.labelNames("id", "name")
-			.help("Stats on world tick times.")
+			.help("Stats on dimension tick times.")
 			.create();
 	}
 
@@ -93,23 +100,40 @@ public class MinecraftCollector extends Collector implements Collector.Describab
 	public List<MetricFamilySamples> collect() {
 		// Collect metrics.
 		MetricFamilySamples player_list = this.collectPlayerList();
-		List<MetricFamilySamples> server_tick = this.server_tick_seconds.collect();
-		MetricFamilySamples world_chunks_loaded = this.collectWorldChunksLoaded();
-		List<MetricFamilySamples> world_tick = this.world_tick_seconds.collect();
+		List<MetricFamilySamples> server_ticks = this.server_tick_seconds.collect();
+		MetricFamilySamples dim_chunks_loaded = this.collectDimensionChunksLoaded();
+		List<MetricFamilySamples> dim_ticks = this.dim_tick_seconds.collect();
 
 		// Aggregate metrics.
 		ArrayList<MetricFamilySamples> metrics = new ArrayList<>(
 			1 /* player_list */
-			+ server_tick.size()
-			+ 1 /* world_chunks_loaded */
-			+ world_tick.size()
+			+ server_ticks.size()
+			+ 1 /* dim_chunks_loaded */
+			+ dim_ticks.size()
 		);
 		metrics.add(player_list);
-		metrics.addAll(server_tick);
-		metrics.add(world_chunks_loaded);
-		metrics.addAll(world_tick);
+		metrics.addAll(server_ticks);
+		metrics.add(dim_chunks_loaded);
+		metrics.addAll(dim_ticks);
 
 		return metrics;
+	}
+
+	/**
+	 * Get the number of loaded dimension chunks.
+	 *
+	 * @return The dimension chunks loaded metric.
+	 */
+	private GaugeMetricFamily collectDimensionChunksLoaded() {
+		GaugeMetricFamily metric = newDimensionChunksLoadedMetric();
+		for (ServerLevel world : this.mc_server.getAllLevels()) {
+			ResourceKey<Level> dim = world.dimension();
+			String id_str = Integer.toString(getDimensionId(dim));
+			String name = dim.location().getPath();
+			int loaded = world.getChunkSource().getLoadedChunksCount();
+			metric.addMetric(Arrays.asList(id_str, name), loaded);
+		}
+		return metric;
 	}
 
 	/**
@@ -121,25 +145,9 @@ public class MinecraftCollector extends Collector implements Collector.Describab
 		GaugeMetricFamily metric = newPlayerListMetric();
 		for (ServerPlayer player : this.mc_server.getPlayerList().getPlayers()) {
 			GameProfile profile = player.getGameProfile();
-			String id = profile.getId().toString();
+			String id_str = profile.getId().toString();
 			String name = profile.getName();
-			metric.addMetric(Arrays.asList(id, name), 1);
-		}
-		return metric;
-	}
-
-	/**
-	 * Get the number of loaded world chunks.
-	 *
-	 * @return The world chunks loaded metric.
-	 */
-	private GaugeMetricFamily collectWorldChunksLoaded() {
-		GaugeMetricFamily metric = newWorldChunksLoadedMetric();
-		for (ServerLevel world : this.mc_server.getAllLevels()) {
-			String id = Integer.toString(getDimensionId(world.dimension()));
-			String name = world.dimension().location().getPath();
-			int loaded = world.getChunkSource().getLoadedChunksCount();
-			metric.addMetric(Arrays.asList(id, name), loaded);
+			metric.addMetric(Arrays.asList(id_str, name), 1);
 		}
 		return metric;
 	}
@@ -155,8 +163,8 @@ public class MinecraftCollector extends Collector implements Collector.Describab
 		ArrayList<MetricFamilySamples> descs = new ArrayList<>();
 		descs.add(newPlayerListMetric());
 		descs.addAll(this.server_tick_seconds.describe());
-		descs.add(newWorldChunksLoadedMetric());
-		descs.addAll(this.world_tick_seconds.describe());
+		descs.add(newDimensionChunksLoadedMetric());
+		descs.addAll(this.dim_tick_seconds.describe());
 		return descs;
 	}
 
@@ -166,23 +174,37 @@ public class MinecraftCollector extends Collector implements Collector.Describab
 	 * With the new version of Minecraft, v16, a dimension no longer has an id.
 	 * However, to keep backward compatibility with older versions of the
 	 * exporter, we need this method. Vanilla dimensions use fixed id values (-1,
-	 * 0, 1), and the id of a custom dimension is now calculated from the world
-	 * name.
+	 * 0, 1), and the id of a custom dimension is now calculated from the
+	 * dimension name.
 	 *
-	 * @param dimension The dimension.
+	 * @param dim The dimension.
 	 */
-	private static int getDimensionId(ResourceKey<Level> dimension) {
-		if (dimension.equals(Level.OVERWORLD)) {
+	private static int getDimensionId(ResourceKey<Level> dim) {
+		if (dim.equals(Level.OVERWORLD)) {
 			return 0;
-		} else if (dimension.equals(Level.END)) {
+		} else if (dim.equals(Level.END)) {
 			return 1;
-		} else if (dimension.equals(Level.NETHER)) {
+		} else if (dim.equals(Level.NETHER)) {
 			return -1;
 		} else {
-			String dimensionName = dimension.location().getPath();
-			return dimensionName.hashCode();
+			String name = dim.location().getPath();
+			return name.hashCode();
 		}
 	}
+
+	/**
+	 * Create a new metric for the dimension chunks loaded.
+	 *
+	 * @return The dimension chunks loaded metric.
+	 */
+	private static GaugeMetricFamily newDimensionChunksLoadedMetric() {
+		return new GaugeMetricFamily(
+			"mc_dimension_chunks_loaded",
+			"The number of loaded dimension chunks.",
+			Arrays.asList("id", "name")
+		);
+	}
+
 	/**
 	 * Create a new metric for the player list.
 	 *
@@ -197,16 +219,22 @@ public class MinecraftCollector extends Collector implements Collector.Describab
 	}
 
 	/**
-	 * Create a new metric for the world chunks loaded.
+	 * Record when a dimension tick begins.
 	 *
-	 * @return The world chunks loaded metric.
+	 * @param dim The dimension.
 	 */
-	private static GaugeMetricFamily newWorldChunksLoadedMetric() {
-		return new GaugeMetricFamily(
-			"mc_world_chunks_loaded",
-			"The number of loaded world chunks.",
-			Arrays.asList("id", "name")
-		);
+	public void startDimensionTick(ResourceKey<Level> dim) {
+		String name = dim.location().getPath();
+		if (this.dim_tick_timer != null) {
+			throw new IllegalStateException(
+				"Dimension " + name + " tick started before stopping previous tick for "
+				+ "dimension " + this.dim_tick_name + "."
+			);
+		}
+
+		String id_str = Integer.toString(getDimensionId(dim));
+		this.dim_tick_name = name;
+		this.dim_tick_timer = this.dim_tick_seconds.labels(id_str, name).startTimer();
 	}
 
 	/**
@@ -214,23 +242,35 @@ public class MinecraftCollector extends Collector implements Collector.Describab
 	 */
 	public void startServerTick() {
 		if (this.server_tick_timer != null) {
-			throw new IllegalStateException("Server tick started before stopping previous tick.");
+			throw new IllegalStateException(
+				"Server tick started before stopping previous tick."
+			);
 		}
+
 		this.server_tick_timer = this.server_tick_seconds.startTimer();
 	}
 
 	/**
-	 * Record when a world tick begins.
+	 * Record when a dimension tick finishes.
 	 *
-	 * @param world The world.
+	 * @param dim The dimension.
 	 */
-	public void startWorldTick(Level world) {
-		if (this.world_tick_timer != null) {
-			throw new IllegalStateException("World tick started before stopping previous tick.");
+	public void stopDimensionTick(ResourceKey<Level> dim) {
+		String name = dim.location().getPath();
+		if (this.dim_tick_timer == null) {
+			throw new IllegalStateException(
+				"Dimension " + name + " tick stopped without an active tick."
+			);
+		} else if (!name.equals(this.dim_tick_name)) {
+			throw new IllegalStateException(
+				"Dimension " + name + " tick stopped while in an active tick for "
+				+ "dimension " + this.dim_tick_name + "."
+			);
 		}
-		String id = Integer.toString(getDimensionId(world.dimension()));
-		String name = world.dimension().location().getPath();
-		this.world_tick_timer = this.world_tick_seconds.labels(id, name).startTimer();
+
+		this.dim_tick_timer.observeDuration();
+		this.dim_tick_timer = null;
+		this.dim_tick_name = null;
 	}
 
 	/**
@@ -238,22 +278,12 @@ public class MinecraftCollector extends Collector implements Collector.Describab
 	 */
 	public void stopServerTick() {
 		if (this.server_tick_timer == null) {
-			throw new IllegalStateException("Server tick stopped without an active tick.");
+			throw new IllegalStateException(
+				"Server tick stopped without an active tick."
+			);
 		}
+
 		this.server_tick_timer.observeDuration();
 		this.server_tick_timer = null;
-	}
-
-	/**
-	 * Record when a world tick finishes.
-	 *
-	 * @param world The world.
-	 */
-	public void stopWorldTick(Level world) {
-		if (this.world_tick_timer == null) {
-			throw new IllegalStateException("World tick stopped without an active tick.");
-		}
-		this.world_tick_timer.observeDuration();
-		this.world_tick_timer = null;
 	}
 }